+++ README.md
@@ -442,3 +442,4 @@ Clear.args = {
 -   Добавлен репорт unit-тесттов в файл main.yml
 -   Добавлен [index.html файл](/reports/index.html) для перехода на страницу с отчетами скриншотных и unit тестов
 -   Обновлен кастомный [eslint-plugin](https://www.npmjs.com/package/eslint-plugin-babun4ek-fsd-plugin/v/0.0.15?activeTab=readme). Рефакторинг
+-   Обновлен redux-toolkit, react-redux до актуальных версий. Обновлен storybook, исправлен loki
