+++ package.json
@@ -24,8 +24,8 @@
         "test:ui:html": "npx reg-cli --from .loki/report.json --report .loki/report.html",
         "analyze:dev": "webpack --env mode=development analyze=true",
         "analyze:prod": "webpack --env mode=production analyze=true",
-        "storybook": "start-storybook -p 6006 -c ./config/storybook",
-        "storybook:build": "build-storybook -c ./config/storybook",
+        "storybook": "storybook dev -p 6006 -c ./config/storybook",
+        "storybook:build": "storybook build -c ./config/storybook",
         "create:slice": "node ./scripts/createSlice/index.js",
         "postinstall": "node ./scripts/clear-cache.js",
         "prepare": "husky install"
@@ -44,14 +44,14 @@
         "@babel/preset-env": "^7.16.11",
         "@babel/preset-react": "^7.16.7",
         "@pmmmwh/react-refresh-webpack-plugin": "^0.5.5",
-        "@storybook/addon-actions": "^6.4.19",
-        "@storybook/addon-essentials": "^6.4.19",
-        "@storybook/addon-interactions": "^6.4.19",
-        "@storybook/addon-links": "^6.4.19",
-        "@storybook/builder-webpack5": "^6.4.19",
-        "@storybook/manager-webpack5": "^6.4.19",
-        "@storybook/react": "^6.4.19",
-        "@storybook/testing-library": "^0.0.9",
+        "@storybook/addon-actions": "^7.4.0",
+        "@storybook/addon-essentials": "^7.4.0",
+        "@storybook/addon-interactions": "^7.4.0",
+        "@storybook/addon-links": "^7.4.0",
+        "@storybook/cli": "^7.4.0",
+        "@storybook/react": "^7.4.0",
+        "@storybook/react-webpack5": "^7.4.0",
+        "@storybook/testing-library": "^0.2.0",
         "@svgr/webpack": "^6.2.1",
         "@testing-library/jest-dom": "^5.16.2",
         "@testing-library/react": "^14.0.0",
@@ -83,7 +83,7 @@
         "eslint-plugin-jsx-a11y": "^6.5.1",
         "eslint-plugin-react": "^7.29.2",
         "eslint-plugin-react-hooks": "^4.3.0",
-        "eslint-plugin-storybook": "^0.6.4",
+        "eslint-plugin-storybook": "^0.6.13",
         "eslint-plugin-unused-imports": "^3.0.0",
         "file-loader": "^6.2.0",
         "fork-ts-checker-webpack-plugin": "^8.0.0",
@@ -102,7 +102,7 @@
         "reg-cli": "^0.17.6",
         "sass": "^1.49.9",
         "sass-loader": "^12.6.0",
-        "storybook-addon-mock": "^2.4.1",
+        "storybook-addon-mock": "^4.2.1",
         "storybook-addon-themes": "^6.1.0",
         "style-loader": "^3.3.1",
         "stylelint": "^14.5.3",
@@ -123,7 +123,7 @@
     "dependencies": {
         "@headlessui/react": "^1.6.6",
         "@react-spring/web": "^9.5.2",
-        "@reduxjs/toolkit": "^1.8.0",
+        "@reduxjs/toolkit": "^1.9.5",
         "@use-gesture/react": "^10.2.19",
         "axios": "^0.26.1",
         "concurrently": "^7.0.0",
@@ -133,10 +133,11 @@
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
         "react-i18next": "^13.0.3",
-        "react-redux": "^7.2.6",
+        "react-redux": "^8.1.2",
         "react-router-dom": "^6.2.1"
     },
     "loki": {
+        "chromeSelector": "#storybook-root > *",
         "configurations": {
             "chrome.laptop": {
                 "target": "chrome.docker",
